pyproject.toml
[tool.poetry]
name = "fastapi-mongodb"
version = "0.1.0"
description = ""
authors = ["pashamakhilkumarreddy <pashamakhilkumarreddy@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
httpx = "^0.27.0"
uvicorn = "^0.30.1"
pymongo = "^4.7.3"
pillow = "^10.3.0"
pydantic-settings = "^2.3.3"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.13.2"
pip-audit = "^2.7.3"
pre-commit = "^3.7.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
setuptools = "^70.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.blue]
line-length = 120
